ortho htrack: remove unnecessary cast
`track_no` implicitly coerces to `double` here.

+++ lib/ortho/ortho.c
@@ -1081,7 +1081,7 @@ vtrack (segment* seg, maze* m)
 
 static double htrack(segment *seg, maze *m) {
   channel* chp = chanSearch(m->hchans, seg);
-  double f = 1.0 - (double)seg->track_no / ((double)seg_list_size(&chp->seg_list) + 1);
+  double f = 1.0 - seg->track_no / ((double)seg_list_size(&chp->seg_list) + 1);
   double lo = chp->cp->bb.LL.y;
   double hi = chp->cp->bb.UR.y;
   return round(lo + f * (hi - lo));
